Major refactor of the Navigation module
Slowly working through the core classes, removing any of the hidden
foo._bar hidden methods and variables. The Navigation module now is
totally self-contained, and exposes only a few methods, hiding away
lots of the hackery I used to get it to work originally.

The animation code has also been tweaked so that the animation speed
will vary depending on the power of the processor. I'm not getting
accurate rendering times for each individual paint event, but rendering
seems to be pretty smooth across all browsers.