Fix up the expansion and collapsing of subgroups in the navigation
Expansion and collapsing of subgroups on the navigation pane
didn't work. This was because the events were not being fired
for the sub-layers (the flag to consume the events was being
passed). This change unsets that flag for the sub-layers, as
well as fixing the getGroup method so that it can handle being
passed a layer. Further, a test to check the size of a group
has been changed from group.length to group.size(). This
code was never being hit previously. Now it will be.


Former-commit-id: ee101be5fccba892b73c7c734b936ac2e679a5fd