Fix up the momentum action on touch devices
Momentum has been a bit buggy on touch devices
this has been caused by the momentum callback
firing off when another drag action has
already been started. To solve this, we
manually tweak the momentum parameters given
that dragging is happening. This allows for the
current x value plus the momentum delta to
be calculated.