Fix up the end of touch events for button zooming
When you ended a zoom on a touch device where you had
used the buttons, it wouldn't fire the completion of the
zoom gesture. This fixes up the logic so that if you
do have a "center" zoomCenter value, it remembers that
you haven't set it, and doesn't wait for the gesture
to end.


Former-commit-id: 0856d302010ba432260c548816eedd5bf6cf2212