Client use of clustalrunner as well as re-align
If you set the endpoint for the client to a local server, instead
of the EBI, it will retrieve the locally cached data.

There is also an extra method for doing a re-alignment of data,
so that if you have a sequence that you aligned, and it has gaps
in it, the sequences can be munged so that the sequence you
wan't doesn't have any gaps, while the others have gaps or have points
where insertions need to be made.


Former-commit-id: a8c317a7f1c4ee3de3b34d791edbbd885b8a80dd