Allow transactions to be queued.
If we have multiple transactions that may occur at the same time
we need a way to sync the actual action of the operations.

We queue up the callbacks to notify the end of the transactions
and then fire them when the reference count for the transactions
is down to zero.


Former-commit-id: a2c8dba39a4b1826d616b4e56fe7f0b306da9e66